+++ ci-test/test_fft+gradient.cpp
@@ -20,7 +20,7 @@ int main()
     /*                    this part is a copy of test_fft                          */
     /*           but care, here x is the continuous like variable                  */
     /*-----------------------------------------------------------------------------*/
-    size_t Npoints = 1000;
+    int Npoints = 1000;
     mdarray<std::complex<double>,2> Array_x({1,Npoints});
     mdarray<std::complex<double>,2> Array_k({1,Npoints});
     std::vector<int> dimensions = {int(Npoints)};
